chore(schematics): auto-move color/gradient utils from cdk to @taiga-ui/editor in v5
Address review: these utils were relocated to @taiga-ui/editor, not removed.
Convert the removed-warnings (incl. the earlier tuiHexToRgb warning) into
IDENTIFIERS_TO_REPLACE package moves so imports are rewritten automatically.
