+++ projects/cdk/schematics/ng-update/v5/steps/constants/migration-warnings.ts
@@ -226,32 +226,6 @@ export const MIGRATION_WARNINGS: MigrationWarning[] = [
             'TUI_ALERT_POSITION has been removed. Alert position is now part of the notification options: use the block and inline properties (e.g. via tuiNotificationOptionsProvider from @taiga-ui/core) instead.',
     },
     {
-        name: 'tuiHexToRgb',
-        moduleSpecifier: '@taiga-ui/cdk',
-        message:
-            'tuiHexToRgb has been removed without a direct replacement. Convert a HEX color to RGB manually where you used it.',
-    },
-    ...[
-        'tuiGetGradientData',
-        'tuiHexToRGBA',
-        'tuiHsvToRgb',
-        'tuiIsValidHex',
-        'tuiIsValidRgba',
-        'tuiParseColor',
-        'tuiParseGradient',
-        'tuiParseHex',
-        'tuiRgbaToHex',
-        'tuiRgbToHex',
-        'tuiRgbToHsv',
-        'tuiToGradient',
-        'TuiGradientDirection',
-        'TuiParsedGradient',
-    ].map((name) => ({
-        name,
-        moduleSpecifier: '@taiga-ui/cdk',
-        message: `${name} has been removed without a direct replacement. Perform the color/gradient calculation manually where you used it.`,
-    })),
-    {
         name: 'TUI_TEXTFIELD_APPEARANCE',
         moduleSpecifier: '@taiga-ui/legacy',
         message:
